Move pyocaction code from init to pyocclient_wrapper
