+++ python/smashbox/utilities/__init__.py
@@ -288,62 +288,6 @@ def webdav_mkcol(path,silent=False):
         out = "> /dev/null 2>&1"
     runcmd('curl -k %s -X MKCOL %s %s'%(config.get('curl_opts',''),oc_webdav_url(remote_folder=path),out))
 
-def pyocaction(username, password, async, method, *args, **kwargs):
-    """
-    Run an action using the pyocclient. This method will create an ownCloud client and run
-    the method. A new client will be created each time.
-
-    :param username: username for the client (for authentication purposes)
-    :param password: password for the client (for authentication purposes)
-    :param async: run the method async? True = async, False = sync
-    :param method: the name of the method to be run from the owncloud client (check pyocclient
-    Client object to know what methods you can use
-    :param *args: arguments that will be passed to the method
-    :param **kwargs: arguments that will be passed to the method. The extra keyword
-    'pyocactiondebug' will enable debug in the Client object but won't be passed to the method
-
-    :return: a tuple containing a Thread object and a Queue (to get the result of the called
-    method once it finish) if the async param is set to True, and the result of the method
-     if is set to False.
-
-    examples:
-    `pyocaction(user, pass, False, 'put_file_contents', '/path/to/file', 'file content')`
-    `thread = pyocaction(user, pass, True, 'get_file', '/bigfile')`
-
-    """
-    import owncloud
-    oc_url = pyocclient_basic_url()
-    if kwargs.get('pyocactiondebug'):
-        client = owncloud.Client(oc_url, debug=True)
-        del kwargs['pyocactiondebug']
-    else:
-        client = owncloud.Client(oc_url)
-    client.login(username, password)
-
-    caller = getattr(client, method)
-    if async:
-        import threading
-        import Queue
-
-        def caller_wrapper(method, q, *args, **kwargs):
-            q.put(method(*args, **kwargs))
-
-        result_queue = Queue.Queue()
-        caller_wrapper_args = (caller, result_queue,) + args
-        thread = threading.Thread(target=caller_wrapper, args=caller_wrapper_args, kwargs=kwargs)
-        thread.start()
-        return (thread, result_queue)
-    else:
-        return caller(*args, **kwargs)
-
-def pyocclient_basic_url():
-    """
-    Return the url to be used by pyocclient-related functions / classes
-    """
-    oc_protocol = 'https' if config.oc_ssl_enabled else 'http'
-    oc_url = oc_protocol + '://' + config.oc_server + '/' + config.oc_root
-    return oc_url
-
 # #### SHELL COMMANDS AND TIME FUNCTIONS
 
 def runcmd(cmd,ignore_exitcode=False,echo=True,allow_stderr=True,shell=True,log_warning=True):
